Revert "Move contracts into api-contract"
This reverts commit af0e8cf07af1bc8f710442d7302a95c165e8c97b.
