--- packages/api-contract/README.md
@@ -1,13 +1,0 @@
-# @polkadot/api-contract
-
-Interfaces to allow for the encoding and decoding of Substrate contract ABIs.
-
-```js
-import { Abi } from '@polkadot/api-contract';
-
-const abi = new Abi(<...JSON ABI...>);
-
-api.tx.contract
-  .call(<contract addr>, <value>, <max gas>, abi.messages.<method name>(<...params...>))
-  .signAndSend(<keyring pair>, (result: SubmittableResult) => { ... });
-```
