--- packages/api-contract/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@polkadot/api-contract",
-  "version": "0.81.0-beta.17",
-  "description": "Interfaces for interacting with contracts and contract ABIs",
-  "main": "index.js",
-  "keywords": [
-    "Polkadot",
-    "ABI"
-  ],
-  "author": "Jaco Greeff <jacogr@gmail.com>",
-  "maintainers": [
-    "Jaco Greeff <jacogr@gmail.com>"
-  ],
-  "contributors": [],
-  "license": "Apache-2.0",
-  "publishConfig": {
-    "access": "public",
-    "registry": "https://registry.npmjs.org"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/polkadot-js/api.git"
-  },
-  "bugs": {
-    "url": "https://github.com/polkadot-js/api/issues"
-  },
-  "homepage": "https://github.com/polkadot-js/api/tree/master/packages/api-contract#readme",
-  "dependencies": {
-    "@babel/runtime": "^7.4.5",
-    "@polkadot/types": "^0.81.0-beta.17"
-  }
-}
