--- packages/api-contract/src/index.ts
@@ -1,6 +1,0 @@
-// Copyright 2017-2019 @polkadot/api-contract authors & contributors
-// This software may be modified and distributed under the terms
-// of the Apache-2.0 license. See the LICENSE file for details.
-
-export { default as Abi } from './Abi';
-export { default as Contract } from './Contract';
