+++ packages/api/test/data/erc20.wasm
