+++ packages/api/test/data/flipper-pruned.wasm
