+++ packages/api/test/data/shared_vec-pruned.wasm
