+++ packages/types/src/index.ts
@@ -9,5 +9,6 @@ import { assertSingletonPackage } from '@polkadot/util';
 
 assertSingletonPackage('@polkadot/types');
 
+export { default as ContractAbi } from './ContractAbi';
 export * from './codec';
 export * from './index.types';
